Move test to different classes for smaller mini ozone cluster.
